move remaining functions to _files.py
